Added partners component
