+++ content/footer.json
@@ -2,7 +2,7 @@
   "sectionOne": [
     {
       "_block": "columnOne",
-      "logo": "assets/logo.png",
+      "logo": "assets/components/logo.png",
       "description": "Our team offers the most up-to-date, sustainable manufacturing solutions. We only source materials from tried and trusted suppliers."
     },
     {
@@ -53,7 +53,7 @@
       "socialLinks": [
         {
           "_block": "social",
-          "icon": "assets/icons8-facebook-50.png",
+          "icon": "assets/components/icons8-facebook-50.png",
           "href": "https://novelaquatech.com/"
         }
       ]
