+++ src/app/laboratory-equipment/page.tsx
@@ -5,6 +5,7 @@ import { ResearchPage } from "@/types/research-page";
 import { getPageBySlug } from "@/lib/api";
 import Link from "next/link";
 import FooterSection from "../_components/footer/footer";
+import PartnersCarousel from "../_components/partners/partner";
 
 export default function LaboratoryEquipmentPage() {
   const labEquipment: ResearchPage = getPageBySlug("research-page.json");
@@ -74,6 +75,7 @@ export default function LaboratoryEquipmentPage() {
           ))}
         </div>
       </div>
+      <PartnersCarousel/>
       <FooterSection/>
     </main>
   );
