+++ src/app/software-engineering/page.tsx
@@ -1,6 +1,7 @@
 import { getPageBySlug } from "../../lib/api";
 import Navbar from "../_components/navbar/navbar";
-
+import FooterSection from "../_components/footer/footer";
+import PartnersCarousel from "../_components/partners/partner";
 import React from "react";
 import SectionSelector from "../_components/sectionSelector";
 
@@ -11,7 +12,7 @@ import {
   IndustriesItem,
   PointsItem,
 } from "@/types/software-page";
-import FooterSection from "../_components/footer/footer";
+
 
 export default function Index() {
   const homePageContent: SoftwarePage = getPageBySlug("software-page.json");
@@ -41,6 +42,7 @@ export default function Index() {
           return <CallToActionSection key={section.id} section={section} />;
         return null;
       })}
+      <PartnersCarousel/>
       <FooterSection/>
     </main>
   );
